+++ event-store-management/event-store-management-core/src/main/java/uk/gov/justice/services/eventstore/management/replay/process/ReplayEventToComponentRunner.java
@@ -1,6 +1,7 @@
 package uk.gov.justice.services.eventstore.management.replay.process;
 
-import java.util.Optional;
+import uk.gov.justice.services.common.util.UtcClock;
+
 import java.util.UUID;
 
 import javax.inject.Inject;
@@ -13,16 +14,8 @@ public class ReplayEventToComponentRunner {
     @Inject
     private ReplayEventToEventListenerProcessorBean replayEventToEventListenerProcessorBean;
 
-    public void run(final UUID commandId, final UUID commandRuntimeId, final String componentName, final Optional<String> eventSourceNameOptional) {
-
-        final String eventSourceName;
-        if (eventSourceNameOptional.isPresent()) {
-            eventSourceName = eventSourceNameFinder.ensureEventSourceNameExistsInRegistry(
-                    eventSourceNameOptional.get(),
-                    componentName);
-        } else {
-            eventSourceName = eventSourceNameFinder.getEventSourceNameOf(componentName);
-        }
+    public void run(final UUID commandId, final UUID commandRuntimeId, final String componentName) {
+        final String eventSourceName = eventSourceNameFinder.getEventSourceNameOf(componentName);
 
         final ReplayEventContext replayEventContext = new ReplayEventContext(commandId,
                 commandRuntimeId, eventSourceName, componentName);